prepare framework structre
