+++ Cargo.toml
@@ -56,6 +56,8 @@ members = [
   "nomos-services/wallet",
   "nomos-tracing",
   "nomos-utils",
+  "testing-framework/configs",
+  "testing-framework/core",
   "testnet/cfgsync",
   "tests",
   "utxotree",
@@ -79,7 +81,7 @@ broadcast-service            = { default-features = false, path = "./nomos-servi
 bundler                      = { default-features = false, path = "./nomos-bundler" }
 cfgsync                      = { default-features = false, path = "./testnet/cfgsync" }
 chain-common                 = { default-features = false, path = "./nomos-services/chain/chain-common" }
-chain-leader                 = { default-features = false, path = "./nomos-services/chain/chain-leader" }
+chain-leader                 = { default-features = false, path = "./nomos-services/chain/chain-leader", features = ["pol-dev-mode"] }
 chain-service                = { default-features = false, path = "./nomos-services/chain/chain-service" }
 circuits-prover              = { default-features = false, path = "./zk/circuits/prover" }
 circuits-utils               = { default-features = false, path = "./zk/circuits/utils" }
@@ -89,6 +91,7 @@ cryptarchia-engine           = { default-features = false, path = "./consensus/c
 cryptarchia-sync             = { default-features = false, path = "./consensus/cryptarchia-sync" }
 executor-http-client         = { default-features = false, path = "./nodes/nomos-executor/http-client" }
 groth16                      = { default-features = false, path = "./zk/groth16" }
+integration-configs          = { default-features = false, path = "testing-framework/configs" }
 key-management-system        = { default-features = false, path = "./nomos-services/key-management-system" }
 key-management-system-macros = { default-features = false, path = "./nomos-services/key-management-system-macros" }
 kzgrs                        = { default-features = false, path = "./nomos-da/kzgrs" }
@@ -128,6 +131,7 @@ poq                          = { default-features = false, path = "zk/proofs/poq
 poseidon2                    = { default-features = false, path = "./zk/poseidon2" }
 services-utils               = { default-features = false, path = "./nomos-services/utils" }
 subnetworks-assignations     = { default-features = false, path = "./nomos-da/network/subnetworks-assignations" }
+testing-framework-core       = { default-features = false, path = "./testing-framework/core" }
 tests                        = { default-features = false, path = "./tests" }
 tx-service                   = { default-features = false, path = "./nomos-services/tx-service" }
 utxotree                     = { default-features = false, path = "./utxotree" }
@@ -147,6 +151,7 @@ overwatch         = { default-features = false, git = "https://github.com/logos-
 overwatch-derive  = { default-features = false, git = "https://github.com/logos-co/Overwatch", rev = "f5a9902" }
 rand              = { default-features = false, version = "0.8" }
 reqwest           = { default-features = false, version = "0.12" }
+serde             = { default-features = true, version = "1.0", features = ["derive"] }
 serde_json        = { default-features = false, version = "1.0" }
 serde_with        = { default-features = false, version = "3.14.0" }
 serde_yaml        = { default-features = false, version = "0.9.33" }
