+++ testing-framework/configs/src/nodes/mod.rs
@@ -1,0 +1,2 @@
+pub mod executor;
+pub mod validator;
