+++ tests/Cargo.toml
@@ -1,69 +1,34 @@
 [package]
-categories  = { workspace = true }
-description = { workspace = true }
-edition     = { workspace = true }
-keywords    = { workspace = true }
-license     = { workspace = true }
-name        = "tests"
-publish     = false
-readme      = { workspace = true }
-repository  = { workspace = true }
-version     = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+name    = "tests"
+publish = false
+version = "0.1.0"
 
 [lints]
 workspace = true
 
+[features]
+default = []
+
 [dependencies]
-async-trait              = "0.1"
-blst                     = { version = "0.3.11" }
-broadcast-service        = { workspace = true }
-chain-leader             = { workspace = true }
 chain-service            = { workspace = true }
 common-http-client       = { workspace = true }
 cryptarchia-engine       = { features = ["serde"], workspace = true }
-cryptarchia-sync         = { workspace = true }
 ed25519-dalek            = { version = "2.2.0", features = ["rand_core", "serde"] }
 executor-http-client     = { workspace = true }
 futures                  = { default-features = false, version = "0.3" }
 futures-util             = { default-features = false, version = "0.3.31" }
-groth16                  = { workspace = true }
-hex                      = { default-features = false, version = "0.4.3" }
-key-management-system    = { workspace = true }
 kzgrs-backend            = { workspace = true }
-nomos-api                = { workspace = true }
-nomos-blend-message      = { workspace = true }
-nomos-blend-service      = { features = ["libp2p"], workspace = true }
 nomos-core               = { workspace = true }
-nomos-da-dispersal       = { workspace = true }
-nomos-da-network-core    = { workspace = true }
-nomos-da-network-service = { workspace = true }
-nomos-da-sampling        = { workspace = true }
-nomos-da-verifier        = { workspace = true }
-nomos-executor           = { default-features = false, features = ["testing", "tracing"], workspace = true }
-nomos-http-api-common    = { workspace = true }
-nomos-ledger             = { features = ["serde"], workspace = true }
 nomos-libp2p             = { workspace = true }
-nomos-network            = { features = ["libp2p"], workspace = true }
-nomos-node               = { default-features = false, features = ["testing"], workspace = true }
-nomos-sdp                = { workspace = true }
-nomos-time               = { workspace = true }
-nomos-tracing            = { workspace = true }
-nomos-tracing-service    = { workspace = true }
 nomos-utils              = { workspace = true }
-nomos-wallet             = { workspace = true }
-num-bigint               = { default-features = false, version = "0.4" }
 rand                     = { workspace = true }
-reqwest                  = { features = ["json"], workspace = true }
-serde_json               = "1.0"
-serde_yaml               = "0.9"
+reqwest                  = { workspace = true, features = ["json"] }
 serial_test              = "3.1.1"
 subnetworks-assignations = { workspace = true }
-tempfile                 = { workspace = true }
-time                     = { default-features = false, version = "0.3" }
+testing-framework-core   = { workspace = true, default-features = false }
 tokio                    = "1"
-tracing                  = { workspace = true }
-tx-service               = { features = ["libp2p", "mock"], workspace = true }
-zksign                   = { workspace = true }
 
 [[test]]
 name = "test_cryptarchia_happy_path"
@@ -100,7 +65,3 @@ path = "src/tests/membership/disperse.rs"
 [[test]]
 name = "test_da_historic_sampling"
 path = "src/tests/da/historic_sampling.rs"
-
-[[test]]
-name = "test_mantle_sdp_ops"
-path = "src/tests/mantle/sdp_ops.rs"
