+++ tests/src/nodes.rs
@@ -1,0 +1,1 @@
+pub use testing_framework_core::nodes::*;
