+++ tests/src/topology.rs
@@ -1,0 +1,3 @@
+pub use testing_framework_core::topology::{
+    self, Topology, TopologyConfig, configs, create_kms_configs,
+};
